move MekSearchFilter
